Report on symbols that have changed between kernel versions
Module.symvers includes a list of the exported kernel symbols AND their crc checksums
so by figuring out which kernel the kmod was compiled against and comparing its
crc numbers against the crc numbers of the targert kernel we can figure out which kernel
symbols have changed between the versions and report them.

This commit rewrites large chunks of the code to support this.